pythonWrapper/requirements.txt
attrdict==2.0.0
bitstring==3.1.5
certifi==2018.4.16
chardet==3.0.4
cytoolz==0.9.0.1
eth-abi==1.1.1
eth-account==0.2.3
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-utils==1.0.3
hexbytes==0.1.0
idna==2.7
lru-dict==1.1.6
parsimonious==0.8.0
py-solc==3.1.0
pycryptodome==3.6.3
requests==2.19.1
rlp==1.0.1
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
urllib3==1.23
web3==4.4.1
websockets==5.0.1
